Refactor pybind11 code to fix configuration bug
